+++ pkg/lib/headers.go
@@ -1,0 +1,5 @@
+package lib
+
+func MakeAuthHeader(token string) string {
+	return "Bearer " + token
+}
